Small clean up of completion code
