+++ src/visitors/semantic/mod.rs
@@ -2,20 +2,15 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 use crate::protocol::properties::Position;
-use crate::protocol::responses::{
-    CompletionItem, CompletionItemKind, InsertTextFormat,
-};
-use crate::stdlib::Completable;
-
-// pub mod walk;
 
 use flux::semantic::nodes::*;
-use flux::semantic::types::MonoType;
 use flux::semantic::walk::{self, Node, Visitor};
 
+mod completion;
 mod symbols;
 pub mod utils;
 
+pub use completion::CompletableFinderVisitor;
 pub use symbols::SymbolsVisitor;
 
 fn contains_position(node: Rc<Node<'_>>, pos: Position) -> bool {
@@ -213,242 +208,6 @@ impl<'a> Visitor<'a> for ImportFinderVisitor {
             (*state).imports.push(import.path.value.clone());
         }
 
-        true
-    }
-}
-
-#[derive(Default)]
-pub struct CompletableFinderState {
-    pub completables: Vec<Rc<dyn Completable>>,
-}
-
-pub struct CompletableFinderVisitor {
-    pub pos: Position,
-    pub state: Rc<RefCell<CompletableFinderState>>,
-}
-
-impl CompletableFinderVisitor {
-    pub fn new(pos: Position) -> Self {
-        CompletableFinderVisitor {
-            state: Rc::new(RefCell::new(
-                CompletableFinderState::default(),
-            )),
-            pos,
-        }
-    }
-}
-
-fn follow_function_pipes(c: Box<CallExpr>) -> MonoType {
-    if let Some(p) = c.pipe {
-        if let Expression::Call(call) = p {
-            return follow_function_pipes(call);
-        }
-    }
-
-    c.typ
-}
-
-impl<'a> Visitor<'a> for CompletableFinderVisitor {
-    fn visit(&mut self, node: Rc<Node<'a>>) -> bool {
-        let mut state = self.state.borrow_mut();
-        let loc = node.loc();
-        let pos = self.pos.clone();
-
-        if loc.start.line > pos.line + 1
-            || (loc.start.line == pos.line + 1
-                && loc.start.column > pos.character + 1)
-        {
-            return true;
-        }
-
-        if let Node::VariableAssgn(assgn) = node.as_ref() {
-            let id = assgn.id.clone();
-            match assgn.init.clone() {
-                Expression::Integer(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name,
-                        var_type: VarType::Int,
-                    }))
-                }
-                Expression::Boolean(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name,
-                        var_type: VarType::Bool,
-                    }))
-                }
-                Expression::Float(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name.clone(),
-                        var_type: VarType::Float,
-                    }))
-                }
-                Expression::StringLit(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name.clone(),
-                        var_type: VarType::String,
-                    }))
-                }
-                Expression::Array(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name.clone(),
-                        var_type: VarType::Array,
-                    }))
-                }
-                Expression::Regexp(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name.clone(),
-                        var_type: VarType::Regexp,
-                    }))
-                }
-                Expression::Duration(_) => {
-                    (*state).completables.push(Rc::new(VarResult {
-                        name: id.name.clone(),
-                        var_type: VarType::Duration,
-                    }))
-                }
-                Expression::Call(c) => {
-                    let result_type = follow_function_pipes(c);
-                    match result_type {
-                        MonoType::Int => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Int,
-                                },
-                            ));
-                        }
-                        MonoType::Float => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Float,
-                                },
-                            ));
-                        }
-                        MonoType::Bool => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Bool,
-                                },
-                            ));
-                        }
-                        MonoType::Arr(_) => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Array,
-                                },
-                            ));
-                        }
-                        MonoType::Duration => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Duration,
-                                },
-                            ));
-                        }
-                        MonoType::Row(_) => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Row,
-                                },
-                            ));
-                        }
-                        MonoType::String => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::String,
-                                },
-                            ));
-                        }
-                        MonoType::Uint => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Uint,
-                                },
-                            ));
-                        }
-                        MonoType::Time => {
-                            (*state).completables.push(Rc::new(
-                                VarResult {
-                                    name: id.name.clone(),
-                                    var_type: VarType::Time,
-                                },
-                            ));
-                        }
-                        _ => {}
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        true
-    }
-}
-
-#[derive(Clone)]
-enum VarType {
-    Int,
-    String,
-    Array,
-    Float,
-    Bool,
-    Duration,
-    Regexp,
-    Row,
-    Uint,
-    Time,
-}
-
-#[derive(Clone)]
-struct VarResult {
-    pub name: String,
-    pub var_type: VarType,
-}
-
-impl VarResult {
-    pub fn detail(&self) -> String {
-        match self.var_type {
-            VarType::Array => "Array".to_string(),
-            VarType::Bool => "Boolean".to_string(),
-            VarType::Duration => "Duration".to_string(),
-            VarType::Float => "Float".to_string(),
-            VarType::Int => "Integer".to_string(),
-            VarType::Regexp => "Regular Expression".to_string(),
-            VarType::String => "String".to_string(),
-            VarType::Row => "Row".to_string(),
-            VarType::Time => "Time".to_string(),
-            VarType::Uint => "Unsigned Integer".to_string(),
-        }
-    }
-}
-
-impl Completable for VarResult {
-    fn completion_item(&self) -> CompletionItem {
-        CompletionItem {
-            label: format!("{} ({})", self.name, "self".to_string()),
-            additional_text_edits: None,
-            commit_characters: None,
-            deprecated: false,
-            detail: Some(self.detail()),
-            documentation: Some("from self".to_string()),
-            filter_text: Some(self.name.clone()),
-            insert_text: Some(self.name.clone()),
-            insert_text_format: InsertTextFormat::PlainText,
-            kind: Some(CompletionItemKind::Variable),
-            preselect: None,
-            sort_text: None,
-            text_edit: None,
-        }
-    }
-
-    fn matches(&self, _text: String, _imports: Vec<String>) -> bool {
         true
     }
 }
